Move classes to their own files. Store electricity data in InfluxDB.
