+++ app/requirements.txt
@@ -1,4 +1,3 @@
 requests >= 2.22.0
-influxdb == 5.2.3
+influxdb-client == 1.40.0
 Click == 7.0
-maya == 0.6.1
