+++ docker-compose.yml
@@ -6,8 +6,8 @@ services:
     - DOCKER_INFLUXDB_INIT_MODE=setup
     - DOCKER_INFLUXDB_INIT_USERNAME=admin
     - DOCKER_INFLUXDB_INIT_PASSWORD=admin_password
-    - DOCKER_INFLUXDB_INIT_ORG=metrics
-    - DOCKER_INFLUXDB_INIT_BUCKET=octograph
+    - DOCKER_INFLUXDB_INIT_ORG=monitoring
+    - DOCKER_INFLUXDB_INIT_BUCKET=octopus-energy-metrics
     - DOCKER_INFLUXDB_INIT_RETENTION=104w
     - DOCKER_INFLUXDB_INIT_ADMIN_TOKEN=secret_admin_token
     ports:
